Move ICRSOrbitAdapter from space_orbit.py into its test file
ICRSOrbitAdapter has no caller anywhere in pycbc itself (checked: only
its own definition and the __all__ export referenced it) -- its only
real use is enabling this test file's independent cross-validation
strategy: wrapping a real, already-constructed lisaorbits object
(analytic, e.g. lisaorbits.KeplerianOrbits, or real-ESA-file-backed,
e.g. lisaorbits.OEMOrbits) directly, in memory, so pycbc's own orbit
math can be compared against it without a round trip through a file
just to pick up the ICRS -> ecliptic conversion.

Since it's test-only, moved the class itself into
test/test_coordinates_space_orbit.py, right before TestICRSOrbitAdapter
(the 9 call sites elsewhere in that file -- TestKeplerianOrbits,
TestICRSOrbitAdapter, TestOptionalESAOemOrbitFiles -- updated to use
the now-local class instead of space_orbit.ICRSOrbitAdapter). Removed
it from space_orbit.py and from pycbc.coordinates' __all__.

The one thing that does stay in space_orbit.py, unchanged: the private
_icrs_to_ecliptic_rotation_matrix() function ICRSOrbitAdapter wraps --
that's genuinely shared production code, also called directly by
NumericOrbits.from_oem_files/from_lisaorbits_file and
rotation_matrix_ssb_to_geo. Updated two docstrings that cross-referenced
"ICRSOrbitAdapter" by name to reference that function directly instead,
since the class they pointed at is no longer in this module.

Verified: full test_coordinates_space_orbit.py (85 tests, including all
ICRSOrbitAdapter-dependent ones: TestICRSOrbitAdapter's 5,
TestOptionalLisaorbitsDuckTyping's Keplerian cross-check, and both
TestOptionalESAOemOrbitFiles real-ESA-data tests) plus
test_coordinates_space.py (2) and test_transforms.py (4) pass with no
regressions. qlty check clean on all three changed files.
